refactor(deploy): Fully split sequential deployment code into packages
This make the separation of current and old code clear on a cmd/ level in the
same way as was done on pkg/ level already.
