+++ cmd/monaco/integrationtest/v2/invalid_manifests_produce_user_errors_e2e_test.go
@@ -88,7 +88,7 @@ func TestNonExistentProjectInManifestReturnsError(t *testing.T) {
 	cmd.SetArgs([]string{"deploy", "--verbose", "--dry-run", manifest})
 	err := cmd.Execute()
 
-	assert.ErrorContains(t, err, "error while loading projects")
+	assert.ErrorContains(t, err, "failed to load projects")
 
 	runLog := strings.ToLower(logOutput.String())
 	expectedErrorLog := "filepath `this_does_not_exist` does not exist"
